Ensure host is privileged to read its resource
The tests are updated to run against the latest 0.7.0 release of v5 Conjur, in
which hosts/users do not own their own resources by default.

Since the Service Broker relies on accessing the resource of the host in its
Conjur configuration to determine platform annotations, we have updated it
to return a 403 response if the host does not have read access to its resource.

We have also updated the health check to verify the host identity has this privilege, and added a cucumber test for the new functionality (for v5 only).