Update env variable presence check to account for empty string
When an environment variable was configured, but to an empty string, the service broker would use that empty value, rather than fallback to the desired default behavior. This updates the check to account for empty strings, and updates the tests to better reflect how these values are presented in the PCF tile.