refactor and add service binding for space host identity
Rather than add additional conditional logic within the single `ServiceBinding` class, I refactored this to put each condition (based on Conjur version and enabled configuration) in its on class. The controller now loads the correct ServiceBinding logic for the current configuration and uses it to create and delete the service binding.