Speed up builds by running all tests and checks in parallel
Since sequential testing of the docker image and unit tests don't
ovrlap, their testing stages can run in parallel to speed up cycles on
this repo. Also, since we may want the `zip` available even if image
scans fail, move that step to a conditional one for the test stage.