Use Gemfile groups to manage production vs dev dependencies
This change uses Gemfile groups to manage dependencies in terms
of which dependencies will be included in:
- Production images
- Development images
- Test images

Before this change, dependencies were managed and separated by including a
special "GEMFILE TRIM MARKER" text block in the Gemfile to delineate
production dependencies from development and test dependencies.
This trim marker was then used by a `sed` command to create temporary
copies of the Gemfile with portions trimmed as required for
production images.