Remove buildpack health check source from images
The dev / test images include the Go-based buildpack health check source right
now, but the build script builds the Go binary and adds it to the "bin/" dir
before the Docker images are built, and the Go binary is really all that's
required to be included in the Docker images. This commit removes the Go
source from the Docker images.