Fixes unknown yaml2 revision build errors
Currently, Jenkins builds are failing for the build of the Buildpack Health
Check executable, with the following error:

```
[2021-10-15T13:16:37.534Z] go: finding gopkg.in/yaml.v2 v2.2.1
[2021-10-15T13:16:37.795Z] go: gopkg.in/yaml.v2@v2.2.1: unknown revision v2.2.1
[2021-10-15T13:16:39.721Z] go: error loading module requirements
```

This appears to be related to a known bug that used to happen for old
versions of Go:
   https://github.com/golang/go/issues/24099
that has since been fixed.

The fix is to use Golang version 1.15 instead of 1.12 for building the
Buildpack Health Check executable.